# integration-framework/moirai-zeus/src/main/resources/static/swagger/apis/openapi.yaml
openapi: 3.0.0
info:
  title: Moirai Zeus API
  description: This is the API for the Zeus(Ta3) system's access to the subprojects (e.g. Clotho),
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /users/authenticate:
    post:
      tags:
        - Users
      summary: Takes username and password and provides authentication.
      description: authenticates user
      operationId: authenticate
      requestBody:
        description: Authenticate user with password
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/zeusLoginRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Invalid credentials.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/account:
    get:
      tags:
        - Users
      summary: Checks for a jwt token. If so, returns the account.
      description: Gets user information.
      operationId: getUserAccount
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/jwtUser'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Invalid credentials.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
...
  /users/settings:
    put:
      tags:
       - Users
      summary: Updates a subset of fields in the jwtUser object known as Settings
      operationId: updateUserSettings
      requestBody:
        description: Change a user's settings
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/settingsDto'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Invalid username or password
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict. Account name taken.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/forgotUsername:
    post:
      tags:
        - Users
      summary: Takes an email and to that email send the coresponding username.
      description: forgot username
      operationId: forgotUsername
      requestBody:
        description: The email address of the user.
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        204:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/forgotPassword:
    post:
      tags:
        - Users
      summary: Takes a username and emailed the associated account a access code.
      description: forgot password
      operationId: forgotPassword
      requestBody:
        description: The username address of the account.
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        204:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/resetPassword:
    post:
      tags:
        - Users
      summary: Takes an accessCode and grants a jwt to resetPassword for the associated account.
      description: reset password
      operationId: resetPassword
      requestBody:
        description: The accessCode.
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/setPassword:
    post:
      tags:
        - Users
      summary: Takes the new password and checks the jwt for access to set the password.
      description: set password
      operationId: setPassword
      requestBody:
        description: The newpassword.
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/contactrequest:
    get:
      tags:
        - Users
      summary: Allows admin to view contact request on status
      description: List Contact Requests for a given status
      operationId: listContactRequestsByStatus
      parameters:
        - name: status
          in: query
          description: Contents to retrieve by status
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/contactRequest'
        404:
          description: Not found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Invalid credentials.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/contact:
    post:
      tags:
        - Users
      summary: Sends contact request to admins
      description: Contact request
      operationId: addContactRequest
      requestBody:
        description: Contact information
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/contactRequestDto'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /users/registerNewTeamMember:
    post:
      tags:
        - Users
      summary: Creates User account for existing team
      description: Creates User account for existing team
      operationId: registerUserToTeam
      requestBody:
        description: Registers user with password
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/registrationDto'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: Conflict. Account name taken.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    zeusLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "moirai_group"
        password:
          type: string
          example: "AVerySecurePassword"
    registrationDto:
      type: object
      required:
        - username
        - password
        - teamName
        - performerGroup
        - emailAddress
      properties:
        username:
          type: string
          example: "persongroup"
        password:
          type: string
        teamName:
          type: string
          example: "Next Century"
        performerGroup:
          type: string
          example: "TA1"
        emailAddress:
          type: string
          example: "generic.email@host.com"
        timezone:
          type: string
          example: "GMT-05:00"
...
    settingsDto:
      type: object
      required:
        - id
        - username
        - password
        - teamName
        - performerGroup
        - emailAddress
      properties:
        id:
          type: string
        username:
          type: string
          example: "persongroup"
        currentPassword:
          type: string
        password:
          type: string
        teamName:
          type: string
          example: "Next Century"
        performerGroup:
          type: string
          example: "TA1"
        emailAddress:
          type: string
          example: "generic.email@host.com"
        timezone:
          type: string
          example: "GMT-05:00"
        profilePictureURL:
          type: string
        darkMode:
          type: boolean
          example: true
...
    jwtUser:
      x-jpaEntity: true
      type: object
      required:
        - username
        - password
        - emailAddress
        - teamName
        - performerGroup
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        active:
          type: boolean
          x-jpaColOverride:
            overrideValue: "columnDefinition = \"tinyint(1) default 1\""
        username:
          type: string
          example: "moirai_group"
        password:
          type: string
          example: "AVerySecurePassword"
        passwordExpiration:
          type: string
          format: date
          example: "2020-01-01"
        teamName:
          type: string
          example: "Next Century"
        performerGroup:
          type: string
          example: "TA1"
        emailAddress:
          type: string
          example: "generic.email@host.com"
        profilePictureURL:
          type: string
        darkMode:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/jwtRole'
          x-lazyManyToMany: true
          x-joinTable:
            joinValue: name = "jwt_user_jwt_role",
              joinColumns = { @JoinColumn(name = "jwt_user_id") },
              inverseJoinColumns = { @JoinColumn(name = "jwt_role_id") }
        timezone:
          type: string
          example: "GMT-05:00"
        services:
          type: array
          items:
            $ref: '#/components/schemas/service'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-cascade-all: true
    stringResponse:
      type: object
      properties:
        value:
          type: string
          example: "Success"
    stringRequest:
      type: object
      properties:
        value:
          type: string
          example: "Sample_text"
    jwtPermission:
      type: object
      x-jpaEntity: true
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "can_provision_graph_db"
        description:
          type: string
          example: "Has access to the provisioning service for graph databases."
    jwtRole:
      type: object
      x-jpaEntity: true
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "admin"
        description:
          type: string
          example: "Administrative User"
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/jwtPermission'
          x-lazyManyToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "jwt_permission_jwt_role",
              joinColumns = { @JoinColumn(name = "jwt_role_id") },
              inverseJoinColumns = { @JoinColumn(name = "jwt_permission_id") }
    service:
      type: object
      x-jpaEntity: true
      required:
        - name
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "service1"
        type:
          type: string
          example: "clotho"
        subtype:
          type: string
          example: "Neo4j"
        status:
          type: string
        dateReceived:
          type: string
          format: date
          example: "2020-01-01"
        teamName:
          type: string
        access:
          type: string
          example: "Private"
        details:
          type: string
        error:
          type: string
          example: "Error Msg"
          x-jpaTransient: true
        awsInstances:
          type: array
          items:
            $ref: '#/components/schemas/serviceAwsInstance'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_aws_instance",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_aws_instance_id") }
          x-cascade-all: true
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/serviceEndpoint'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_endpoint",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_endpoint_id") }
          x-cascade-all: true
        downloads:
          type: array
          items:
            $ref: '#/components/schemas/serviceDownload'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_download",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_download_id") }
          x-cascade-all: true
    serviceAwsInstance:
      type: object
      x-jpaEntity: true
      required:
        - instanceId
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        instanceId:
          type: string
          example: "i-011d626d652684d5a4"
    serviceEndpoint:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Clotho Endpoint"
        uri:
          type: string
          example: "ec2-18-232-154-122.compute-1.amazonaws.com:8008"
    serviceDownload:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Python Connector"
        uri:
          type: string
          example: "s3://kairos-service-artifacts/connectors/next_century/service_a/connector.py"
    contactRequestDto:
      type: object
      x-jpaEntity: false
      required:
        - email
        - topic
        - message
      properties:
        requestor:
          type: string
          example: "generic.email@host.com"
        topic:
          type: string
          example: "Services"
        message:
          type: string
          example: "Sending you a message as a test."
    contactRequest:
      type: object
      x-jpaEntity: true
      required:
        - requestor
        - topic
        - datetime
        - status
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        requestor:
          type: string
          example: "someone@email.com"
        topic:
          type: string
          example: "Access"
        datetime:
          type: string
          format: date-time
          example: "12449123"
        status:
          type: string
          example: "open"
        messages:
          type: array
          items:
            $ref: '#/components/schemas/contactRequestMessage'
          x-lazyOneToMany: true
          x-cascade-all: true
    contactRequestMessage:
      type: object
      x-jpaEntity: true
      required:
        - messageId
        - datetime
        - sender
        - text
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        datetime:
          type: string
          format: date-time
          example: "12449123"
        sender:
          type: string
          example: "someone@email.com"
        text:
          type: string
          example: "This is the message"
...